+++ lib/active_record/framing/railtie.rb
@@ -4,7 +4,6 @@ require 'active_record/framing/query_methods'
 require 'active_record/framing/spawn_methods'
 require 'active_record/framing/attribute_methods'
 require 'active_record/framing/relation'
-require 'active_record/framing/associations/join_dependency'
 
 module ActiveRecord::Framing
   class Railtie < Rails::Railtie
@@ -14,7 +13,6 @@ module ActiveRecord::Framing
         ::ActiveRecord::Relation.prepend(ActiveRecord::Framing::Relation)
         ::ActiveRecord::Relation.include(ActiveRecord::Framing::QueryMethods)
         ::ActiveRecord::Relation.prepend(ActiveRecord::Framing::SpawnMethods)
-        ::ActiveRecord::Associations::JoinDependency.prepend(ActiveRecord::Framing::Associations::JoinDependency)
       end
     end
   end