+++ lib/deleted_at/core_ext/kernel.rb
@@ -1,5 +1,0 @@
-module Kernel
-  def tap?
-    yield(self) if block_given? && self
-  end
-end