+++ lib/deleted_at/registry.rb
@@ -1,7 +1,0 @@
-require 'active_support/per_thread_registry'
-
-module DeletedAt
-  class QueryRegistry # :nodoc:
-    extend ActiveSupport::PerThreadRegistry
-  end
-end