+++ spec/support/rails/app/models/animals/cat.rb
@@ -1,5 +1,0 @@
-module Animals
-  class Cat < ::ActiveRecord::Base
-
-  end
-end