+++ spec/support/rails/app/models/animals/mean_cat.rb
@@ -1,5 +1,0 @@
-module Animals
-  class MeanCat < Cat
-
-  end
-end